adding class User that encapsulates methods that are used by student , professor , TA
