+++ Notifications/src/Course.java
@@ -2,9 +2,9 @@ import java.util.ArrayList;
 
 import gateways.EmailGateway;
 import Messages.TaskAddedEmailMessage;
-import Users.Professor;
-import Users.Student;
-import Users.TA;
+import users.Professor;
+import users.Student;
+import users.TA;
 
 public class Course {
