+++ Notifications/src/users/Student.java
@@ -1,0 +1,26 @@
+package users;
+
+public class Student extends User {
+	int id;
+
+
+	public Student(String name, int id, String email, String phoneNumber) {
+		super();
+		this.name = name;
+		this.id= id;
+		this.email = email;
+		this.phoneNumber = phoneNumber;
+	}
+
+	public int getId() {
+		return id;
+	}
+
+	public void setId(int id) {
+		this.id = id;
+	}
+
+
+
+
+}
