+++ Notifications/src/users/TA.java
@@ -1,0 +1,24 @@
+package users;
+
+public class TA extends User {
+	String department;
+
+	public TA(String name, String department, String email, String phoneNumber) {
+		super();
+		this.name = name;
+		this.department = department;
+		this.email = email;
+		this.phoneNumber = phoneNumber;
+	}
+
+	public String getDepartment() {
+		return department;
+	}
+
+	public void setDepartment(String department) {
+		this.department = department;
+	}
+
+
+
+}
